% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-jyear.R
\name{convert_jyear}
\alias{convert_jyear}
\title{Convert Japanese imperial year to Anno Domini}
\usage{
convert_jyear(jyear, legacy = getOption("zipangu.convert_jyear.legacy", TRUE))
}
\arguments{
\item{jyear}{Japanese imperial year (jyear). Kanji or Roman character}

\item{legacy}{A logical to switch converter. If \code{TRUE} (the default), use the
legacy converter based on \code{lubridate}. If \code{FALSE}, use the 'ICU' implementation
via \code{stringi::stri_datetime_parse()} with the \code{ja-JP-u-ca-japanese} locale.
The default can also be controlled globally by
\code{options(zipangu.convert_jyear.legacy = FALSE)}.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}
}
\section{ICU implementation note}{

The non-legacy implementation depends on the system 'ICU' library through
\code{stringi}. On some platforms (notably macOS with system 'ICU'), parsing with
the \code{ja-JP-u-ca-japanese} locale can abort the R process. The legacy path is
therefore used by default. See \url{https://github.com/uribo/zipangu/issues/60}.
}

\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{convert_jyear("R1")
#> [1] 2019
convert_jyear("Heisei2")
#> [1] 1990
convert_jyear("\\u5e73\\u6210\\u5143\\u5e74")
#> [1] 1989
convert_jyear(c("\\u662d\\u548c10\\u5e74", "\\u5e73\\u621014\\u5e74"))
#> [1] 1935 2002
convert_jyear(kansuji2arabic_all("\\u5e73\\u6210\\u4e09\\u5e74"))
#> [1] 1991
}\if{html}{\out{</div>}}
}

