% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_prefecture.R
\name{convert_prefecture}
\alias{convert_prefecture}
\title{Convert prefecture names to roman or kanji}
\usage{
convert_prefecture(x, to)
}
\arguments{
\item{x}{prefecture name in kanji}

\item{to}{conversion destination}
}
\description{
Convert prefecture names to roman or kanji
}
\examples{
convert_prefecture(c("tokyo-to", "osaka", "ALL"), to="kanji")
convert_prefecture(
  c("\u6771\u4eac", "\u5927\u962a\u5e9c",
  "\u5317\u6d77\u9053", "\u5168\u56fd"),
  to = "roman")
}
