% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_prefecture_from_kana.R
\name{convert_prefecture_from_kana}
\alias{convert_prefecture_from_kana}
\title{Convert prefecture names from kana}
\usage{
convert_prefecture_from_kana(x)
}
\arguments{
\item{x}{prefecture name in kana}
}
\description{
Convert prefecture names from kana
}
\examples{
convert_prefecture_from_kana(c("\u3068\u3046\u304d\u3087\u3046\u3068"))
convert_prefecture_from_kana(c("\u30c8\u30a6\u30ad\u30e7\u30a6\u30c8", "\u30ad\u30e7\u30a6\u30c8"))
convert_prefecture_from_kana(c("\u30c8\u30a6\u30ad\u30e7\u30a6", "\u304a\u304a\u3055\u304b"))

}
