% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_prefecture.R
\name{convert_prefecture_to_kanji}
\alias{convert_prefecture_to_kanji}
\title{Convert prefecture names to kanji}
\usage{
convert_prefecture_to_kanji(x)
}
\arguments{
\item{x}{prefecture name in roman}
}
\description{
Convert prefecture names to kanji
}
