% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_prefecture.R
\name{convert_prefecture_to_roman}
\alias{convert_prefecture_to_roman}
\title{Convert prefecture names to roman}
\usage{
convert_prefecture_to_roman(x)
}
\arguments{
\item{x}{prefecture name in kanji}
}
\description{
Convert prefecture names to roman
}
