% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode.R
\name{dl_zipcode_file}
\alias{dl_zipcode_file}
\title{Download a zip-code file}
\usage{
dl_zipcode_file(path, exdir = NULL)
}
\arguments{
\item{path}{local file path or zip file URL}

\item{exdir}{The directory to extract zip file. If \code{NULL}, use temporary folder.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}
}
\examples{
\dontrun{
dl_zipcode_file(path = "https://www.post.japanpost.jp/zipcode/dl/oogaki/zip/02aomori.zip")
dl_zipcode_file("https://www.post.japanpost.jp/zipcode/dl/oogaki/zip/02aomori.zip",
                exdir = getwd())
}
}
