% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jholiday.R
\name{find_date_by_wday}
\alias{find_date_by_wday}
\title{Find out the date of the specific month and weekday}
\usage{
find_date_by_wday(year, month, wday, ordinal)
}
\arguments{
\item{year}{numeric year}

\item{month}{numeric month}

\item{wday}{numeric weekday}

\item{ordinal}{number of week}
}
\value{
a vector of class POSIXct
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
Get the date of the Xth the specific weekday
}
\examples{
find_date_by_wday(2021, 1, 2, 2)
}
