% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode.R
\name{is_japanpost_zippath}
\alias{is_japanpost_zippath}
\title{Check if it is a zip file provided by japanpost}
\usage{
is_japanpost_zippath(url)
}
\arguments{
\item{url}{character.}
}
\value{
A logical vector.
}
\description{
Check if it is a zip file provided by japanpost
}
