% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefecture.R
\name{is_prefecture}
\alias{is_prefecture}
\title{Check correctly prefecture names}
\usage{
is_prefecture(x)
}
\arguments{
\item{x}{Input vector.}
}
\value{
logical
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("stable")}
}
\details{
Check if the string is a prefectural string.
If it contains the name of the prefecture and other
strings (e.g. city name), it returns \code{FALSE}.
}
\examples{
is_prefecture("\u6771\u4eac\u90fd")
is_prefecture(c("\u6771\u4eac", "\u4eac\u90fd", "\u3064\u304f\u3070"))
}
