% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode.R
\name{is_zipcode}
\alias{is_zipcode}
\title{Test zip-code}
\usage{
is_zipcode(x)
}
\arguments{
\item{x}{Zip-code. Number or character. Hyphens may be included,
but the input must contain a 7-character number.}
}
\value{
A logical vector.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\examples{
is_zipcode(7000027)
is_zipcode("700-0027")
}
