% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jholiday.R
\name{jholiday_spec}
\alias{jholiday_spec}
\alias{jholiday}
\title{Public holidays in Japan}
\usage{
jholiday_spec(year, name, lang = "en")

jholiday(year, lang = "en")
}
\arguments{
\item{year}{numeric year and in and after 1949.}

\item{name}{holiday name}

\item{lang}{return holiday names to "en" or "jp".}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\details{
Holiday information refers to data published as of December 21, 2020.
Future holidays are subject to change.
}
\examples{
jholiday_spec(2019, "Sports Day")
jholiday_spec(2021, "Sports Day")
# List of a specific year holidays
jholiday(2021, "en")
}
\references{
Public Holiday Law \url{https://www8.cao.go.jp/chosei/shukujitsu/gaiyou.html},
\url{https://elaws.e-gov.go.jp/document?lawid=323AC1000000178}
}
