% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-jpnprefs.R
\docType{data}
\name{jpnprefs}
\alias{jpnprefs}
\title{Prefectural informations in Japan}
\format{A tibble with 47 rows 5 variables:
\itemize{
\item{jis_code: jis code}
\item{prefecture_kanji: prefecture names}
\item{prefecture: prefecture names}
\item{region: region}
\item{major_island: }
}}
\usage{
jpnprefs
}
\description{
Prefectures dataset.
}
\examples{
jpnprefs
}
\keyword{datasets}
