% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kana.R
\name{kana}
\alias{kana}
\alias{hiragana}
\alias{katakana}
\title{Create kana vector}
\usage{
kana(type, ...)

hiragana(
  core = TRUE,
  dakuon = FALSE,
  handakuon = FALSE,
  kogaki = FALSE,
  historical = FALSE
)

katakana(
  core = TRUE,
  dakuon = FALSE,
  handakuon = FALSE,
  kogaki = FALSE,
  historical = FALSE
)
}
\arguments{
\item{type}{"hiragana" ("hira") or "katakana" ("kata")}

\item{...}{Arguments passed on to \link{hiragana}}

\item{core}{is include core kana characters.}

\item{dakuon}{e.g. ga, gi, gu, ge, go}

\item{handakuon}{e.g. pa, pi, pu, pe, po}

\item{kogaki}{small character}

\item{historical}{old style}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
Generates a vector consisting of the elements of kana.
Options exist for the inclusion of several elements.
}
\examples{
kana(type = "hira", core = TRUE)
kana(type = "hira", core = TRUE, handakuon = TRUE)
}
