% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_prefecture_list.R
\docType{data}
\name{prefecture_list}
\alias{prefecture_list}
\title{List of prefecture names}
\format{
An object of class \code{list} of length 6.
}
\usage{
prefecture_list
}
\description{
List of prefecture names
}
\keyword{datasets}
