% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode.R
\name{read_zipcode}
\alias{read_zipcode}
\title{Read Japan post's zip-code file}
\usage{
read_zipcode(path, type = c("oogaki", "kogaki", "roman", "jigyosyo"))
}
\arguments{
\item{path}{path}

\item{type}{Input file type, one of "oogaki", "kogaki", "roman", "jigyosyo"}
}
\value{
\link[tibble:tibble]{tibble}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\seealso{
\url{https://www.post.japanpost.jp/zipcode/dl/readme.html},
\url{https://www.post.japanpost.jp/zipcode/dl/jigyosyo/readme.html}
}
