% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{separate_address}
\alias{separate_address}
\title{Separate address elements}
\usage{
separate_address(str)
}
\arguments{
\item{str}{Input vector. address strings.}
}
\value{
A list of elements that make up an address.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
Parses and decomposes address string into elements of
prefecture, city, and lower address.
}
\examples{
separate_address("\u5317\u6d77\u9053\u672d\u5e4c\u5e02\u4e2d\u592e\u533a")
}
