% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode.R
\name{zipcode_spacer}
\alias{zipcode_spacer}
\title{Insert and remove zip-code connect character}
\usage{
zipcode_spacer(x, remove = FALSE)
}
\arguments{
\item{x}{Zip-code. Number or character. Hyphens may be included,
but the input must contain a 7-character number.}

\item{remove}{Default is \code{FALSE}. If \code{TRUE}, remove the hyphen.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}
Inserts a hyphen as a delimiter in the given zip-code string.
Or exclude the hyphen.
}
\examples{
zipcode_spacer(7000027)
zipcode_spacer("305-0053")
zipcode_spacer("305-0053", remove = TRUE)
}
