% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-jyear.R
\name{convert_jdate}
\alias{convert_jdate}
\title{Convert Japanese date format to date object}
\usage{
convert_jdate(date)
}
\arguments{
\item{date}{a character object.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}
}
\examples{
convert_jdate("\u4ee4\u548c2\u5e747\u67086\u65e5")
}
