% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-jyear.R
\name{convert_jdate}
\alias{convert_jdate}
\title{Convert Japanese date format to date object}
\usage{
convert_jdate(date, legacy = FALSE)
}
\arguments{
\item{date}{A character object.}

\item{legacy}{A logical to switch converter. If \code{TRUE} supplied, use the legacy
converter instead of 'ICU' implementation.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{convert_jdate("R3/2/27")
#> [1] "2021-02-27"
convert_jdate("\\u4ee4\\u548c2\\u5e747\\u67086\\u65e5")
#> [1] "2020-07-06"
}\if{html}{\out{</div>}}
}

