% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-jyear.R
\name{convert_jyear}
\alias{convert_jyear}
\title{Convert Japanese imperial year to Anno Domini}
\usage{
convert_jyear(jyear, legacy = FALSE)
}
\arguments{
\item{jyear}{Japanese imperial year (jyear). Kanji or Roman character}

\item{legacy}{A logical to switch converter. If \code{TRUE} supplied, use the legacy
converter instead of 'ICU' implementation.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{convert_jyear("R1")
#> [1] 2019
convert_jyear("Heisei2")
#> [1] 1990
convert_jyear("\\u5e73\\u6210\\u5143\\u5e74")
#> [1] 1989
convert_jyear(c("\\u662d\\u548c10\\u5e74", "\\u5e73\\u621014\\u5e74"))
#> [1] 1935 2002
convert_jyear(kansuji2arabic_all("\\u5e73\\u6210\\u4e09\\u5e74"))
#> [1] 1991
}\if{html}{\out{</div>}}
}

