% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jholiday.R
\name{is_jholiday}
\alias{is_jholiday}
\title{Is x a public holidays in Japan?}
\usage{
is_jholiday(date)
}
\arguments{
\item{date}{a vector of \link{POSIXt}, numeric or character objects}
}
\value{
TRUE if x is a public holidays in Japan, FALSE otherwise.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
Whether it is a holiday defined by Japanese law (enacted in 1948)
}
\details{
Holiday information refers to data published as of December 21, 2020.
Future holidays are subject to change.
}
\section{Examples}{
\if{html}{\out{<div class="r">}}\preformatted{is_jholiday("2021-01-01")
}\if{html}{\out{</div>}}\preformatted{## [1] TRUE
}\if{html}{\out{<div class="r">}}\preformatted{is_jholiday("2018-12-23")
}\if{html}{\out{</div>}}\preformatted{## [1] TRUE
}\if{html}{\out{<div class="r">}}\preformatted{is_jholiday("2019-12-23")
}\if{html}{\out{</div>}}\preformatted{## [1] FALSE
}
}

